+++ Sources/HIGThemesContract/HIGTheme.swift
@@ -31,6 +31,7 @@ public protocol HIGTheme: Sendable {
     var card: any HIGCardTokens { get }
     var panel: any HIGPanelTokens { get }
     var counter: any HIGCounterTokens { get }
+    var widget: any HIGWidgetTokens { get }
     var dataTable: any HIGDataTableTokens { get }
     var breadcrumb: any HIGBreadcrumbTokens { get }
     var pageHeader: any HIGPageHeaderTokens { get }